+++ lib/stream/passthrough.js
@@ -19,32 +19,23 @@
 // OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE
 // USE OR OTHER DEALINGS IN THE SOFTWARE.
 
+// a passthrough stream.
+// basically just the most minimal sort of Transform stream.
+// Every written chunk gets output as-is.
 
-var common = require('../common.js');
-var R = require('stream/readable');
-var assert = require('assert');
+module.exports = PassThrough;
 
+var Transform = require('stream/transform');
 var util = require('util');
-var EE = require('events').EventEmitter;
+util.inherits(PassThrough, Transform);
 
-var ondataCalled = 0;
+function PassThrough(options) {
+  if (!(this instanceof PassThrough))
+    return new PassThrough(options);
 
-function TestReader() {
-  R.apply(this);
-  this._buffer = new Buffer(100);
-  this._buffer.fill('x');
-
-  this.on('data', function() {
-    ondataCalled++;
-  });
+  Transform.call(this, options);
 }
 
-util.inherits(TestReader, R);
-
-TestReader.prototype._read = function(n, cb) {
-  cb(null, this._buffer);
-  this._buffer = new Buffer(0);
+PassThrough.prototype._transform = function(chunk, output, cb) {
+  cb(null, chunk);
 };
-
-var reader = new TestReader();
-assert.equal(ondataCalled, 1);
