+++ lib/stream/writable.js
